Rename dendrites.functional to dendrites.apply_dendrites
