+++ nupic/research/frameworks/dendrites/apply_dendrites.py
