+++ nupic/research/frameworks/dendrites/dendritic_layers.py
@@ -25,7 +25,7 @@ linear layer with the output from a set of dendritic segments.
 """
 import abc
 
-import nupic.research.frameworks.dendrites.functional as dendrite_fxns
+import nupic.research.frameworks.dendrites.apply_dendrites as dendrite_fxns
 from nupic.research.frameworks.dendrites import DendriteSegments
 from nupic.torch.modules.sparse_weights import SparseWeights, SparseWeights2d
